get rid of message_type.h struct
